viewer/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiofiles==23.2.1
python-multipart==0.0.6

# Authentication & Security
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
pydantic[email]==2.5.0
pydantic-settings==2.1.0
requests>=2.31.0

# Image Processing
Pillow>=10.0.1
numpy>=1.24.3
psutil>=5.9.8
tifffile>=2023.7.10
zarr>=2.16.0
